+++ components/bsx/Create/CreateCollectionPage.vue
